Implement NodalBCBase class
