+++ framework/include/systems/NonlinearSystemBase.h
@@ -28,7 +28,7 @@ class NodalDamper;
 class GeneralDamper;
 class GeometricSearchData;
 class IntegratedBCBase;
-class NodalBC;
+class NodalBCBase;
 class PresetNodalBC;
 class DGKernel;
 class InterfaceKernel;
@@ -622,7 +622,7 @@ protected:
   ///@{
   /// BoundaryCondition Warhouses
   MooseObjectWarehouse<IntegratedBCBase> _integrated_bcs;
-  MooseObjectWarehouse<NodalBC> _nodal_bcs;
+  MooseObjectWarehouse<NodalBCBase> _nodal_bcs;
   MooseObjectWarehouse<PresetNodalBC> _preset_nodal_bcs;
   ///@}
